Update csv2psql.py - DOES NOT WORK
If I remove "NOT NULL" constraints from the schema, the program works well. Obviously, the schema requires these constraints, so it's not a workable solution.

I went down the rabbit hole of trying to create cascading, named functions to sequently connect the database. Instead, I need to redesign my program in order to more extensibly populate the database.